requirements.txt
PyPDF2
python-dotenv
langchain-openai
openai 
browser_use
playwright